Move common function `active_leaves()` into lib.rs
